finish login page and manage them

+++ Dashboard/views.py
@@ -7,7 +7,9 @@ import sys
 import requests
 from django.core import serializers
 from django.http import HttpResponse,JsonResponse
-from django.shortcuts import render
+from django.shortcuts import render,redirect
+from django.contrib.auth import authenticate, login,logout
+
 from .models import Server,SSWConfig,SBCConfig,RTPConfig,SSW,SBC,RTP
 
 
@@ -15,15 +17,49 @@ from .models import Server,SSWConfig,SBCConfig,RTPConfig,SSW,SBC,RTP
 
 ###################################################################################################
 
+def Login(request):
+    if(request.user.is_authenticated ==False):
 
+        if(request.method=="POST"):
+            try:
+                username=request.POST['username']
+                password=request.POST['password']
+                user = authenticate(username=username, password=password)
+                login(request, user)
+                return redirect('/')
+            except:
+                return redirect('/login/')
+            # return render(request,"Dashboard_Templates/login.html",{"login_status":"not_login"})
+        else:
+            return render(request,"Dashboard_Templates/login.html")
+    else:
+        return redirect('/')
+        
+
+def Logout(request):
+    if(request.user.is_authenticated ):
+
+        if(request.method=="GET"):
+          
+                logout(request)
+
+                return redirect('/login/')
+
+            # return render(request,"Dashboard_Templates/login.html",{"login_status":"not_login"})
+        else:
+            return render(request,"Dashboard_Templates/login.html")
+    else:
+        return redirect('/')
 
 def MainDashboardPage(request):
-    input_file = open ('config/json/rules.json')
-    json_array = json.load(input_file)
-    input_file.close()
+    if(request.user.is_authenticated ):
+        input_file = open ('config/json/rules.json')
+        json_array = json.load(input_file)
+        input_file.close()
 
-    return render(request,"Dashboard_Templates/dash_base.html",{"all":json_array})
-
+        return render(request,"Dashboard_Templates/dash_base.html",{"all":json_array})
+    else:
+        return redirect('/login/')
 
 
 ###################################################################################################
@@ -34,75 +70,77 @@ def MainDashboardPage(request):
 
 
 def MonitoringServer(request,slug):
-    try : 
-        if(slug=="all-service"):
-            main=Server.objects.all()
-            input_file = open ('config/json/rules.json')
-            json_array = json.load(input_file)
-            input_file.close()
-            status={}
-            for server in main:
-                data={'name':server.server_id,'type':server.Type}
-                r=requests.post(url="http://127.0.0.1:5000/check-status",json=data)
-                status[server.name]=r.text
+    if(request.user.is_authenticated ):
+        try : 
+            if(slug=="all-service"):
+                main=Server.objects.all()
+                input_file = open ('config/json/rules.json')
+                json_array = json.load(input_file)
+                input_file.close()
+                status={}
+                for server in main:
+                    data={'name':server.server_id,'type':server.Type}
+                    r=requests.post(url="http://127.0.0.1:5000/check-status",json=data)
+                    status[server.name]=r.text
             
             
-            return render(request,"Dashboard_Templates/datatable.html",{"alldata":main,"type":"all","all":json_array,'status':status})
+                return render(request,"Dashboard_Templates/datatable.html",{"alldata":main,"type":"all","all":json_array,'status':status})
 
 
-        elif(slug=="ssw"):
-            status={}
-            main=Server.objects.filter(Type="ssw")
-            input_file = open ('config/json/rules.json')
-            json_array = json.load(input_file)
-            input_file.close()
+            elif(slug=="ssw"):
+                status={}
+                main=Server.objects.filter(Type="ssw")
+                input_file = open ('config/json/rules.json')
+                json_array = json.load(input_file)
+                input_file.close()
       
             
-            for server in main:
-                data={'name':server.server_id,'type':server.Type}
-                r=requests.post(url="http://127.0.0.1:5000/check-status",json=data)
-                status[server.name]=r.text
+                for server in main:
+                    data={'name':server.server_id,'type':server.Type}
+                    r=requests.post(url="http://127.0.0.1:5000/check-status",json=data)
+                    status[server.name]=r.text
     
-            return render(request,"Dashboard_Templates/datatable.html",{"alldata":main,"type":"ssw","all":json_array,'status':status})
+                return render(request,"Dashboard_Templates/datatable.html",{"alldata":main,"type":"ssw","all":json_array,'status':status})
             
 
-        elif(slug=="sbc"):
-            status={}
-            main=Server.objects.filter(Type="sbc")
-            input_file = open ('config/json/rules.json')
-            json_array = json.load(input_file)
-            input_file.close()
+            elif(slug=="sbc"):
+                status={}
+                main=Server.objects.filter(Type="sbc")
+                input_file = open ('config/json/rules.json')
+                json_array = json.load(input_file)
+                input_file.close()
 
 
 
-            for server in main:
-                data={'name':server.server_id,'type':server.Type}
-                r=requests.post(url="http://127.0.0.1:5000/check-status",json=data)
-                status[server.name]=r.text
-            return render(request,"Dashboard_Templates/datatable.html",{"alldata":main,"type":"sbc","all":json_array,'status':status})
+                for server in main:
+                    data={'name':server.server_id,'type':server.Type}
+                    r=requests.post(url="http://127.0.0.1:5000/check-status",json=data)
+                    status[server.name]=r.text
+                return render(request,"Dashboard_Templates/datatable.html",{"alldata":main,"type":"sbc","all":json_array,'status':status})
 
 
-        elif(slug=="rtp"):
-            status={}
-            main=Server.objects.filter(Type="rtp")
-            input_file = open ('config/json/rules.json')
-            json_array = json.load(input_file)
-            input_file.close()
+            elif(slug=="rtp"):
+                status={}
+                main=Server.objects.filter(Type="rtp")
+                input_file = open ('config/json/rules.json')
+                json_array = json.load(input_file)
+                input_file.close()
 
 
 
-            for server in main:
-                data={'name':server.server_id,'type':server.Type}
-                r=requests.post(url="http://127.0.0.1:5000/check-status",json=data)
-                status[server.name]=r.text
-            return render(request,"Dashboard_Templates/datatable.html",{"alldata":main,"type":"rtp","all":json_array,'status':status})
+                for server in main:
+                    data={'name':server.server_id,'type':server.Type}
+                    r=requests.post(url="http://127.0.0.1:5000/check-status",json=data)
+                    status[server.name]=r.text
+                return render(request,"Dashboard_Templates/datatable.html",{"alldata":main,"type":"rtp","all":json_array,'status':status})
 
 
-        else:
-            return render(request,"Dashboard_Templates/404.html")
-    except:
-        return HttpResponse("error")
-
+            else:
+                return render(request,"Dashboard_Templates/404.html")
+        except:
+            return HttpResponse("error")
+    else:
+        return redirect("/login/")
 
 
 ###################################################################################################
@@ -113,43 +151,47 @@ def MonitoringServer(request,slug):
 
 
 def AddServer(request):
-    try:
-        if(request.POST['type']=="ssw"):
-            Type=str(request.POST['type'])
-            name=str(request.POST['name'])
-            ip=str(request.POST['ip'])
-            server_object=Server(name=name,Type=Type,ip=ip)
-            server_object.save()
-            ssw_object=SSW(server_id=server_object)
-            ssw_object.save()
-            return HttpResponse("ok")
+    if(request.user.is_authenticated ):
+
+        try:
+            if(request.POST['type']=="ssw"):
+                Type=str(request.POST['type'])
+                name=str(request.POST['name'])
+                ip=str(request.POST['ip'])
+                server_object=Server(name=name,Type=Type,ip=ip)
+                server_object.save()
+                ssw_object=SSW(server_id=server_object)
+                ssw_object.save()
+                return HttpResponse("ok")
 
 
-        elif(request.POST['type']=="sbc"):
-            Type=str(request.POST['type'])
-            name=str(request.POST['name'])
-            ip=str(request.POST['ip'])
-            server_object=Server(name=name,Type=Type,ip=ip)
-            server_object.save()
-            sbc_object=SBC(server_id=server_object)
-            sbc_object.save()
-            return HttpResponse("ok")
+            elif(request.POST['type']=="sbc"):
+                Type=str(request.POST['type'])
+                name=str(request.POST['name'])
+                ip=str(request.POST['ip'])
+                server_object=Server(name=name,Type=Type,ip=ip)
+                server_object.save()
+                sbc_object=SBC(server_id=server_object)
+                sbc_object.save()
+                return HttpResponse("ok")
 
 
-        elif(request.POST['type']=="rtp"):
-            Type=str(request.POST['type'])
-            name=str(request.POST['name'])
-            ip=str(request.POST['ip'])
-            server_object=Server(name=name,Type=Type,ip=ip)
-            server_object.save()
-            rtp_object=RTP(server_id=server_object)
-            rtp_object.save()
+            elif(request.POST['type']=="rtp"):
+                Type=str(request.POST['type'])
+                name=str(request.POST['name'])
+                ip=str(request.POST['ip'])
+                server_object=Server(name=name,Type=Type,ip=ip)
+                server_object.save()
+                rtp_object=RTP(server_id=server_object)
+                rtp_object.save()
 
 
-            return HttpResponse("ok")
-        
-    except:
-        return HttpResponse("error")
+                return HttpResponse("ok")
+            
+        except:
+            return HttpResponse("error")
+    else:
+        return redirect("/login/")
 
 
 
@@ -161,18 +203,21 @@ def AddServer(request):
 
 
 def EditServer(request):
-    try:
-        server_id=request.POST['id']    
-        new_name=request.POST['new_name']
-        new_ip=request.POST['new_ip']
+    if(request.user.is_authenticated ):
+        try:
+            server_id=request.POST['id']    
+            new_name=request.POST['new_name']
+            new_ip=request.POST['new_ip']
 
-        edit_model = Server.objects.get(pk=server_id)
-        edit_model.name=new_name
-        edit_model.ip=new_ip
-        edit_model.save()
-        return HttpResponse("ok")
-    except:
-        return HttpResponse("error")
+            edit_model = Server.objects.get(pk=server_id)
+            edit_model.name=new_name
+            edit_model.ip=new_ip
+            edit_model.save()
+            return HttpResponse("ok")
+        except:
+            return HttpResponse("error")
+    else:
+        return redirect("/login/")
 
 ###################################################################################################
 
@@ -183,14 +228,15 @@ def EditServer(request):
 
 
 def DeleteServer(request):
-    
-    try:
-        server_id=request.POST['id']
-        Server.objects.filter(pk=server_id).delete()
-        return HttpResponse("ok")
-    except:
-        return HttpResponse("error")
-
+    if(request.user.is_authenticated ):
+        try:
+            server_id=request.POST['id']
+            Server.objects.filter(pk=server_id).delete()
+            return HttpResponse("ok")
+        except:
+            return HttpResponse("error")
+    else:
+        return redirect("/login/")
 
 
 
@@ -204,17 +250,19 @@ def DeleteServer(request):
 
 
 def start_server_service(request):
-    try:
-        Type=request.POST['type']
-        ip=request.POST['ip']
-        payload={'type':Type}
-        response = requests.post('http://'+ip+':5000/start-server',json=payload)
+    if(request.user.is_authenticated ):
+        try:
+            Type=request.POST['type']
+            ip=request.POST['ip']
+            payload={'type':Type}
+            response = requests.post('http://'+ip+':5000/start-server',json=payload)
 
 
-        return HttpResponse("ok")
-    except:
-        return HttpResponse("fail")
-
+            return HttpResponse("ok")
+        except:
+            return HttpResponse("fail")
+    else:
+        return redirect("/login/")
 ###################################################################################################
 
 ###################################################################################################
@@ -225,16 +273,19 @@ def start_server_service(request):
 
 
 def stop_server_service(request):
-    try:
-        Type=request.POST['type']
-        ip=request.POST['ip']
-        payload={'type':Type}
-        response = requests.post('http://'+ip+':5000/stop-server',json=payload)
+    if(request.user.is_authenticated ):
+        try:
+            Type=request.POST['type']
+            ip=request.POST['ip']
+            payload={'type':Type}
+            response = requests.post('http://'+ip+':5000/stop-server',json=payload)
 
 
-        return HttpResponse("ok")
-    except:
-        return HttpResponse("fail")
+            return HttpResponse("ok")
+        except:
+            return HttpResponse("fail")
+    else:
+        return redirect('/login/')
 ###################################################################################################
 
 ###################################################################################################
@@ -243,17 +294,19 @@ def stop_server_service(request):
 
 
 def start_subservice(request):
-    try:
-        SubServiceName=request.POST['subservice']
-        ip=request.POST['ip']
-        payload={'subservice':SubServiceName}
-        response = requests.post('http://'+ip+':5000/start-subservice',json=payload)
+    if(request.user.is_authenticated ):
+        try:
+            SubServiceName=request.POST['subservice']
+            ip=request.POST['ip']
+            payload={'subservice':SubServiceName}
+            response = requests.post('http://'+ip+':5000/start-subservice',json=payload)
 
 
-        return HttpResponse("ok")
-    except:
-        return HttpResponse("fail")
-
+            return HttpResponse("ok")
+        except:
+            return HttpResponse("fail")
+    else:
+        return redirect("/login/")
 ###################################################################################################
 
 ###################################################################################################
@@ -264,37 +317,41 @@ def start_subservice(request):
 
 
 def stop_subservice(request):
-    try:
-        SubServiceName=request.POST['subservice']
-        ip=request.POST['ip']
-        payload={'subservice':SubServiceName}
-        response = requests.post('http://'+ip+':5000/stop-subservice',json=payload)
+    if(request.user.is_authenticated ):
+        try:
+            SubServiceName=request.POST['subservice']
+            ip=request.POST['ip']
+            payload={'subservice':SubServiceName}
+            response = requests.post('http://'+ip+':5000/stop-subservice',json=payload)
 
 
-        return HttpResponse("ok")
-    except:
-        return HttpResponse("fail")
-
+            return HttpResponse("ok")
+        except:
+            return HttpResponse("fail")
+    else:
+        return redirect("/login/")
 ###################################################################################################
 
 ###################################################################################################
 
 
 def status_subservice(request):
-    try:
-        if(request.method=='GET'):
+    if(request.user.is_authenticated ):
+        try:
+            if(request.method=='GET'):
 
-            ip=request.GET['ip']
-            Type=request.GET['type']
-            payload = {'type':Type}
-            response = requests.get('http://'+ip+':5000/status-subservice', params=payload)
-            data=json.loads(response.content)
-            return JsonResponse(data=data)
-      
+                ip=request.GET['ip']
+                Type=request.GET['type']
+                payload = {'type':Type}
+                response = requests.get('http://'+ip+':5000/status-subservice', params=payload)
+                data=json.loads(response.content)
+                return JsonResponse(data=data)
+        
 
-    except:
-            return HttpResponse("fail")
-
+        except:
+                return HttpResponse("fail")
+    else:
+        return redirect("/login/")
 
 
 ###################################################################################################
